Licensing ITFoundation under the GNU General Public License version 2.
Previous revisions of ITFoundation are provided only for historical purposes,
and the rights to those revisions are reserved by iThink Software. Each revision
is licensed under the terms specified in the file COPYING, if it exists.

Updating headers to remove CVS/Subversion keywords, and modernize copyright
notices to account for new licensing.